Fix Windows 7 32b Compilation from scratch. Add scripts to build Libs on branch.
